home: Navigate to the top, revert landing changes
