+++ src/pages/home/hero.jsx
@@ -4,22 +4,28 @@ import { FiCamera } from "react-icons/fi";
 import { MdOutlineOpenInNew } from "react-icons/md";
 import ScrollView from "../../components/util/ScrollView";
 import { useRef, useState } from "react";
-import { Link } from "react-router";
+import { Link, useNavigate } from "react-router-dom";
 import PropTypes from 'prop-types';
 
 function Hero({ photos }) {
     const scrollContainerRef = useRef(null);
     const [currentImageIndex, setCurrentImageIndex] = useState(0);
+    const navigate = useNavigate();
+
+    const handlePhotoReelsClick = () => {
+        window.scrollTo(0, 0);
+        navigate('/photo-reels');
+    };
 
     return (
         <div className="flex flex-col py-10">
-            <Link to="/photo-reels">
+            <div onClick={handlePhotoReelsClick} className="cursor-pointer">
                 <p className="flex flex-row items-center gap-2 pt-3 
                     font-bold text-[14px] text-tertiary uppercase">
                     <p>Top Rated on <u>Photo Reels</u></p>
                     <MdOutlineOpenInNew size={20} />
                 </p>
-            </Link>
+            </div>
             <p className="font-playfair font-medium text-[48px] leading-[1.1] text-primary md:py-2">
                 Captured by us, celebrated by all.
             </p>
@@ -36,20 +42,22 @@ function Hero({ photos }) {
                                     src={image.image}
                                     alt="club-gallery"
                                     key={index}
-                                    className={`overflow-visible w-full max-h-[500px] rounded-[8px] object-contain transition-all duration-100 ease-in
-                                    ${index == currentImageIndex ? "brightness-100" : "brightness-90 grayscale-[0.25]"}`}
+                                    className={`overflow-visible w-full max-h-[300px] md:max-h-[400px] lg:max-h-[800px] rounded-[8px] object-contain transition-all duration-100 ease-in
+                                    ${index == currentImageIndex ? "brightness-100" : "brightness-80 grayscale-[0.5]"}`}
                                 />
                             ))}
                         </div>
                     </div>
                 </div>
                 {/* Photo Info */}
-                <div className="flex flex-col gap-4 pt-2 md:w-[400px]">
+                <div className="flex flex-col gap-4 pt-2 justify-between md:min-w-[180px] lg:min-w-[220px]">
                     <div>
                         <p className="font-playfair font-medium text-[16px] leading-[1.2] 
-                            text-primary italic">
+                            text-primary italic tracking-wide">
                             {photos[currentImageIndex].title}
                         </p>
+                    </div>
+                    <div>
                         <div className="flex flex-col gap-4 py-4">
                             <p className="flex flex-row items-center gap-3 text-primary">
                                 <IoLocationOutline size={20} />
@@ -70,16 +78,16 @@ function Hero({ photos }) {
                                 </p>
                             </Link>
                         </div>
+                        <ScrollView
+                            speed={1.5}
+                            currentIndex={currentImageIndex}
+                            setCurrentIndex={setCurrentImageIndex}
+                            totalImages={photos.length}
+                            scrollType="single"
+                            containerRef={scrollContainerRef}
+                            centerCalculation={true}
+                        />
                     </div>
-                    <ScrollView
-                        speed={1.5}
-                        currentIndex={currentImageIndex}
-                        setCurrentIndex={setCurrentImageIndex}
-                        totalImages={photos.length}
-                        scrollType="single"
-                        containerRef={scrollContainerRef}
-                        centerCalculation={true}
-                    />
                 </div>
             </div>
         </div>
