+++ src/pages/home/landing.jsx
@@ -2,7 +2,6 @@ import { ReactTyped } from "react-typed";
 import { FiCamera } from "react-icons/fi";
 import Button from "../../components/Button";
 import landingImg from "../../assets/images/landing.png";
-import { IoIosArrowDown } from "react-icons/io";
 import { TbSpeakerphone } from "react-icons/tb";
 import PropTypes from 'prop-types';
 
@@ -30,7 +29,7 @@ function Landing({ onJoin, specialNotices }) {
                     </Button>
                 </div>
                 <div className="mt-8 px-3 py-3 
-                            rounded-full bg-[#f9ffe8]/70 backdrop-blur-sm 
+                            rounded-full bg-white/80 backdrop-blur-sm 
                             text-black text-sm md:text-base font-medium
                             flex items-center gap-1
                             ">
@@ -44,29 +43,6 @@ function Landing({ onJoin, specialNotices }) {
                         loop
                     />
                 </div>
-            </div>
-            <div className="absolute bottom-0 left-0 w-full h-[100px] flex items-center justify-center md:hidden"
-                onClick={() => {
-                    const targetHeight = document.querySelector('.max-w-container').offsetTop;
-                    const startPosition = window.pageYOffset;
-                    const distance = targetHeight - 50 - startPosition;
-                    const duration = 300;
-                    let start = null;
-
-                    const smoothScroll = (timestamp) => {
-                        if (!start) start = timestamp;
-                        const progress = timestamp - start;
-                        const progressPercentage = Math.min(progress / duration, 1);
-                        window.scrollTo(0, startPosition + distance * progressPercentage);
-                        if (progress < duration) {
-                            window.requestAnimationFrame(smoothScroll);
-                        }
-                    };
-
-                    window.requestAnimationFrame(smoothScroll);
-                }}
-            >
-                <span className="text-black text-sm md:text-base font-medium">Scroll down</span> <IoIosArrowDown className="w-5 h-5" />
             </div>
         </div>
     )
